requirements.txt
SoundFile==0.10.3.post1
sacrebleu==1.5.0
tqdm==4.56.2
pydub==0.24.1
Cython==0.29.21
opencv_contrib_python==4.5.1.48
Flask==1.1.2
numpy==1.19.5
hydra_core==1.0.6
omegaconf==2.0.6
regex==2020.11.13
torchaudio==0.7.2
torch==https://download.pytorch.org/whl/cpu/torch-1.0.1.post2-cp37-cp37m-linux_x86_64.whl
matplotlib==3.3.4
requests==2.25.1
dataclasses==0.8
apex==0.9.10dev
boto3==1.17.5
botocore==1.20.5
editdistance==0.5.3
fairscale==0.1.6
faiss==1.5.3
fastBPE==0.1.0
fvcore==0.1.3.post20210204
gossip==2.4.0
h5py==3.1.0
hydra==2.5
indicnlp==0.0.1
nltk==3.5
pandas==1.2.2
Pillow==8.1.0
psutil==5.8.0
pyarrow==3.0.0
pythainlp==2.2.6
PyYAML==5.4.1
sacremoses==0.0.43
sentencepiece==0.1.95
spacy==3.0.1
subword_nmt==0.3.7
tensorboardX==2.1
tornado==6.1
transformers==4.3.2
vizseq==0.1.15
wandb==0.10.18
gunicorn==20.0.4
